benchmark/functions/python/manual/requirements.txt
opentelemetry-api>=1.28.1,<2.0.0
opentelemetry-sdk>=1.28.1,<2.0.0
opentelemetry-exporter-otlp-proto-http>=1.28.1,<2.0.0
opentelemetry-semantic-conventions>=0.49b1,<1.0.0
otlp-stdout-adapter>=0.1.0,<1.0.0 